+++ opensquirrel/passes/router/shortest_path_router.py
@@ -2,9 +2,10 @@ from typing import Any
 
 import networkx as nx
 
+from opensquirrel import Connectivity
 from opensquirrel.ir import IR
 from opensquirrel.passes.router.common import PathFinderType, ProcessSwaps
-from opensquirrel.passes.router.general_router import Connectivity, Router
+from opensquirrel.passes.router.general_router import Router
 
 
 class ShortestPathRouter(Router):